Add a "global" base User type and a Nullable utility
The `UserBase` type is intended to provide a base type for both requests and responses so that we can reduce the amount of "copy-paste" documentation currently in place.

We additionally include a `Nullable<T>` utility type that allows us to easily mark certain types as having fields that can all be the value `null`. This is useful for certain User response shapes.