Push most User object fields to shared interface
The core "User" interface is shared amongst endpoints with "little" differences like "but the fields can be null" or "but the fields are all optional". Fortunately, we can express this in TypeScript using a handful of utility types (and one we've added ourselves). This makes future maintenance of the code much simpler where User fields are universally applicable.